requirements.txt
openai-whisper
setuptools-rust
pydub
moviepy==1.0.3
